+++ changelog.md
@@ -15,10 +15,12 @@ v1.4 (2021-xx-xx):
 * Account for momentum when determining attack range for custom SF_NOJUMPDAMAGE characters
 * Fix nonsense attraction shield target range check
 * Don't extend target distance for rings etc. in special stages
-* Fix targeting invisible things (fixes getting stuck in CBW's Chaos Mode mod)
+* Fix targeting invisible things
 * Allow Tails to fly-attack targets a bit further away, and fly around in goop smarter
 * Fix various longstanding issues with relative height checks in reverse gravity or at scale
 * Try switching targets if we've failed to hit our current target after 3 jumps
+* Attempt to join matches when spectating by pressing fire
+* Support for CBW's Chaos Mode mod - prioritize objectives, ready up by jumping into emblems
 
 v1.3.1 (2021-06-26):
 --------------------